Refactor register component
