+++ components/Register/styles/Register.module.scss
@@ -24,83 +24,7 @@
   box-shadow: 0px 0px 49px -14px rgba(0, 0, 0, 1);
 }
 
-.card__img-section {
-  min-width: 65%;
-  width: 65%;
-  height: 100%;
 
-  position: relative;
-
-  border-right: 1px solid $primary-color2;
-}
-
-.card__img-section::before {
-  content: "";
-  width: 100%;
-  height: 100%;
-
-  position: absolute;
-
-  top: 0px;
-  left: 0px;
-
-  background: rgba(0, 0, 0, 0.75);
-  background: linear-gradient(
-    90deg,
-    rgba(15, 15, 15, 0.6) 0%,
-    rgba(15, 15, 15, 0.2) 0%
-  );
-
-  -webkit-box-shadow: inset 0px 0px 49px -10px rgba(0, 0, 0, 1);
-  -moz-box-shadow: inset 0px 0px 49px -10px rgba(0, 0, 0, 1);
-  box-shadow: inset 0px 0px 49px -10px rgba(0, 0, 0, 1);
-
-  z-index: 1;
-}
-
-.card__img {
-  width: 100%;
-  height: 100%;
-
-  object-fit: cover;
-  display: none;
-  position: absolute;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-@keyframes fadeOut {
-  from {
-    opacity: 1;
-  }
-  to {
-    opacity: 0;
-  }
-}
-
-.card__img.fade-out {
-  display: block;
-  opacity: 0;
-  animation: fadeOut 3s;
-}
-
-.card__img.active {
-  display: block;
-  opacity: 1;
-  animation: fadeIn 3s;
-}
-
-.card__img.init {
-  opacity: 1;
-  display: block;
-}
 
 .card__form-section {
   width: 100%;
@@ -133,104 +57,7 @@
   padding-bottom: 0.3rem;
 }
 
-.form__inputs {
-  padding-left: 0;
 
-  margin-top: 2rem;
-  margin-bottom: 0;
-  width: 100%;
-
-  @include flexVertical;
-  justify-content: center;
-
-  list-style: none;
-}
-
-.form__input {
-  width: fit-content;
-  @include flexVertical;
-  justify-content: center;
-  width: 100%;
-  margin-bottom: 1.5rem;
-}
-
-.form__input:last-of-type {
-  margin-bottom: 0;
-}
-
-.input__label {
-  font-family: $main-font;
-  color: white;
-
-  margin-bottom: 0.5rem;
-  font-size: 1.3rem;
-}
-
-.input__wrap {
-  width: 65%;
-  @include flexHorizontal;
-  justify-content: center;
-
-  position: relative;
-}
-
-.input {
-  height: fit-content;
-  font-size: 1rem;
-  width: 65%;
-
-  border-radius: 100vw;
-  padding-inline: 1rem;
-  padding-block: 0.4rem;
-}
-
-.input__wrap .input {
-  width: 100%;
-}
-
-.input__eye-icon-btn {
-  height: fit-content;
-  width: fit-content;
-  border: none;
-  position: absolute;
-  right: 12px;
-  background-color: transparent;
-  padding: 0;
-
-  display: flex;
-  align-items: center;
-}
-
-.input__eye-icon {
-  height: 22px;
-  width: auto;
-  fill: gray;
-}
-
-.input__eye-icon.active {
-  fill: $primary-color2;
-}
-
-.input__requirements {
-  margin-top: 0.6rem;
-  list-style: none;
-  padding: 0;
-
-  width: 65%;
-}
-
-.input__requirement {
-  font-size: 12px;
-  font-family: $main-font;
-  font-weight: 200;
-  color: white;
-  margin-bottom: 0.3rem;
-  padding-left: 1rem;
-}
-
-.input__requirement:last-of-type {
-    margin-bottom: 0;
-}
 
 .form__footer {
     width: 100%;
@@ -254,9 +81,6 @@
 }
 
 @media (hover: hover) and (pointer: fine) {
-  .input__eye-icon-btn:hover {
-    cursor: pointer;
-  }
 
   .form__submit-btn:hover {
     cursor: pointer;
