+++ components/Register/styles/RegisterInput.module.scss
@@ -1,0 +1,93 @@
+@import "../../../styles/variables.scss";
+
+.form__input {
+  width: fit-content;
+  @include flexVertical;
+  justify-content: center;
+  width: 100%;
+  margin-bottom: 1.5rem;
+}
+
+.form__input:last-of-type {
+  margin-bottom: 0;
+}
+
+.input__label {
+  font-family: $main-font;
+  color: white;
+
+  margin-bottom: 0.5rem;
+  font-size: 1.3rem;
+}
+
+.input__wrap {
+  width: 65%;
+  @include flexHorizontal;
+  justify-content: center;
+
+  position: relative;
+}
+
+.input {
+  height: fit-content;
+  font-size: 1rem;
+  width: 65%;
+
+  border-radius: 100vw;
+  padding-inline: 1rem;
+  padding-block: 0.4rem;
+}
+
+.input__wrap .input {
+  width: 100%;
+}
+
+.input__eye-icon-btn {
+  height: fit-content;
+  width: fit-content;
+  border: none;
+  position: absolute;
+  right: 12px;
+  background-color: transparent;
+  padding: 0;
+
+  display: flex;
+  align-items: center;
+}
+
+.input__eye-icon {
+  height: 22px;
+  width: auto;
+  fill: gray;
+}
+
+.input__eye-icon.active {
+  fill: $primary-color2;
+}
+
+.input__requirements {
+  margin-top: 0.6rem;
+  list-style: none;
+  padding: 0;
+
+  width: 65%;
+}
+
+.input__requirement {
+  font-size: 12px;
+  font-family: $main-font;
+  font-weight: 200;
+  color: white;
+  margin-bottom: 0.3rem;
+  padding-left: 1rem;
+}
+
+.input__requirement:last-of-type {
+  margin-bottom: 0;
+}
+
+@media (hover: hover) and (pointer: fine) {
+  .input__eye-icon-btn:hover {
+    cursor: pointer;
+  }
+}
