+++ components/Register/styles/RegisterInputs.module.scss
@@ -1,0 +1,14 @@
+@import "../../../styles/variables.scss";
+
+.form__inputs {
+    padding-left: 0;
+  
+    margin-top: 2rem;
+    margin-bottom: 0;
+    width: 100%;
+  
+    @include flexVertical;
+    justify-content: center;
+  
+    list-style: none;
+  }
